Moved function to the repo_controller
