+++ app/controllers/github_controller.py
@@ -10,22 +10,4 @@ from app.controllers.auth_controller import get_installation_access_token
 
 github_controller = Blueprint('github', __name__)
 
-def get_branches(owner, repo_name):
-    """Fetches all branches of a repository."""
-    access_token = get_installation_access_token()
 
-    if not access_token:
-        logging.error("Access token is missing. Cannot fetch branches.")
-        return None
-
-    headers = {'Authorization': f'Bearer {access_token}'}
-    api_url = f'https://api.github.com/repos/{owner}/{repo_name}/branches'
-
-    try:
-        response = requests.get(api_url, headers=headers, timeout=60)
-        response.raise_for_status()
-        branches = response.json()
-        return [branch['name'] for branch in branches]
-    except RequestException as e:
-        logging.error(f"Failed to fetch branches for repository {repo_name}: {e}")
-        return None
